+++ FITApp.EmployeesService/Interfaces/IEmployeesService.cs
@@ -16,12 +16,10 @@ namespace FITApp.EmployeesService.Interfaces
         Task<long> UpdateEmployeeEducations(string id, EducationDto educationDto);
         Task<long> UpdateEmployeeAcademicDegrees(string id, AcademicDegreeDto academicDegreeDto);
         Task<long> UpdateEmployeeAcademicRanks(string id, AcademicRankDto academicRankDto);
-        Task <long> UpdateEmployeePhoto(string id, EmployeePhotoUploadDto photoUploadDto);
         Task<long> RemoveEmployeeAcademicRankByIndex(string id, int index);
         Task<long> RemoveEmployeePositionByIndex(string id, int index);
         Task<long> RemoveEmployeeEducationByIndex(string id, int index);
         Task<long> RemoveEmployeeAcademicDegreeByIndex(string id, int index);
-        Task<long> RemoveEmployeePhoto(string id);
         Task<EmployeesPaginationDto> GetEmployeesPagination(int page, int pageSize);
     }
 }
