+++ FITApp.EmployeesService/Interfaces/IPhotoService.cs
@@ -1,0 +1,11 @@
+﻿using FITApp.EmployeesService.Dtos;
+
+namespace FITApp.EmployeesService.Interfaces
+{
+    public interface IPhotoService
+    {
+        Task<long> UpdateEmployeePhoto(string id, EmployeePhotoUploadDto photoUploadDto);
+        Task<long> RemoveEmployeePhoto(string id);
+
+    }
+}
